fix(review): apply additional code review fixes
- executor.py: use RiskDecision.REJECT enum instead of string literal
- models.py + responses.py: Portfolio.total_return/daily_pnl_pct float → Decimal
- portfolio.py: remove float() conversion, use Decimal.quantize(0.01)
- mock.py: quantize _cash after BUY/SELL to 2 decimal places
- test_executor_service.py: assert result.status == OrderStatus.FILLED (was string "filled")
- test_portfolios.py: total_return=Decimal, add 500 error case tests for
  get_portfolio and sync_portfolio service exceptions

+++ app/core/responses.py
@@ -62,13 +62,13 @@ class PortfolioSummaryResponse(CamelModel):
     buying_power: Decimal
     invested_value: Decimal
     market_value: Decimal
-    total_return: float
+    total_return: Decimal
     total_return_amount: Decimal
     unrealized_pnl: Decimal
     unrealized_pnl_pct: Decimal
     realized_pnl: Decimal
     daily_pnl: Decimal
-    daily_pnl_pct: float
+    daily_pnl_pct: Decimal
     position_count: int
     currency: str
